Start the queue swipe on navigate
Waiting for the next document query delayed the view
transition and showed a skeleton first. Let the router
wrap the location commit so the pane swipes immediately.
